feat: v2 poc
