+++ package.json
@@ -22,16 +22,15 @@
     "test": "jest"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.3",
+    "@types/jest": "^27.5.2",
     "@types/lodash": "^4.14.177",
-    "@types/node": "^16.11.11",
-    "jest": "^27.4.3",
-    "ts-jest": "^27.0.7",
-    "ts-node": "^10.1.0",
-    "typescript": "^4.5.2"
+    "@types/node": "^16.18.61",
+    "jest": "^27.5.1",
+    "ts-jest": "^27.1.5",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   },
   "dependencies": {
-    "chalk": "^4.1.2",
-    "lodash": "^4.17.21"
+    "zod": "^3.22.4"
   }
 }
